refactor(auth): 관리자 예외 처리 구조 정리 (#36)
