+++ backend/src/main/java/com/toadzip/backend/admin/controller/AdminAuthenticationController.java
@@ -3,10 +3,10 @@ package com.toadzip.backend.admin.controller;
 import com.toadzip.backend.admin.dto.AdminLoginRequest;
 import com.toadzip.backend.admin.dto.AdminSessionResponse;
 import com.toadzip.backend.admin.dto.CsrfTokenResponse;
+import com.toadzip.backend.admin.exception.InvalidAdminCredentialsException;
 import com.toadzip.backend.admin.service.AdminAuthenticationAuditService;
 import com.toadzip.backend.admin.service.AdminAuthenticationService;
 import com.toadzip.backend.admin.service.AdminLoginAttemptLimiter;
-import com.toadzip.backend.admin.service.InvalidAdminCredentialsException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
