+++ backend/src/main/java/com/toadzip/backend/admin/service/AdminLoginAttemptLimiter.java
@@ -1,5 +1,6 @@
 package com.toadzip.backend.admin.service;
 
+import com.toadzip.backend.admin.exception.AdminLoginAttemptLimitExceededException;
 import java.time.Clock;
 import java.time.Duration;
 import java.time.Instant;
